refactor: admin/auth featureをserver/client/shared構成にリストラクチャ
