+++ admin/src/app/(protected)/layout.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from "react";
 import { Home, User } from "lucide-react";
-import { LogoutButton } from "@/features/auth/components/logout-button";
-import { getCurrentAdmin } from "@/features/auth/lib/auth-server";
+import { LogoutButton } from "@/features/auth/client/components/logout-button";
+import { getCurrentAdmin } from "@/features/auth/server/lib/auth-server";
 import { NavigationLinks } from "./layout/navigation-links";
 
 export default async function MainLayout({
