+++ admin/src/app/login/page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from "react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { LoginForm } from "@/features/auth/components/login-form";
+import { LoginForm } from "@/features/auth/client/components/login-form";
 
 export default function LoginPage() {
   return (
