+++ admin/src/features/bills-edit/actions/update-bill.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { requireAdmin } from "@/features/auth/lib/auth-server";
+import { requireAdmin } from "@/features/auth/server/lib/auth-server";
 import { invalidateWebCache } from "@/lib/utils/cache-invalidation";
 import { type BillUpdateInput, billUpdateSchema } from "../types";
 import { updateBillRecord } from "../repositories/bill-edit-repository";
