+++ admin/src/features/bills/actions/delete-bill.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { revalidatePath } from "next/cache";
-import { requireAdmin } from "@/features/auth/lib/auth-server";
+import { requireAdmin } from "@/features/auth/server/lib/auth-server";
 import { deleteBillById } from "../repositories/bill-repository";
 
 export async function deleteBill(id: string) {
