+++ admin/src/features/diet-sessions/actions/update-diet-session.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { requireAdmin } from "@/features/auth/lib/auth-server";
+import { requireAdmin } from "@/features/auth/server/lib/auth-server";
 import { invalidateWebCache } from "@/lib/utils/cache-invalidation";
 import type { UpdateDietSessionInput } from "../types";
 import { updateDietSessionRecord } from "../repositories/diet-session-repository";
