+++ admin/src/features/interview-config/actions/generate-interview-preview-url.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { requireAdmin } from "@/features/auth/lib/auth-server";
+import { requireAdmin } from "@/features/auth/server/lib/auth-server";
 import { env } from "@/lib/env";
 import { previewTokenService } from "../../bills/services/preview-token-service";
