+++ admin/src/features/interview-config/actions/upsert-interview-config.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { requireAdmin } from "@/features/auth/lib/auth-server";
+import { requireAdmin } from "@/features/auth/server/lib/auth-server";
 import { invalidateWebCache } from "@/lib/utils/cache-invalidation";
 import { type InterviewConfigInput, interviewConfigSchema } from "../types";
 import {
