+++ admin/src/features/tags/actions/create-tag.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { requireAdmin } from "@/features/auth/lib/auth-server";
+import { requireAdmin } from "@/features/auth/server/lib/auth-server";
 import { invalidateWebCache } from "@/lib/utils/cache-invalidation";
 import type { CreateTagInput } from "../types";
 import { createTagRecord } from "../repositories/tag-repository";
