+++ admin/src/features/tags/actions/update-tag.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { requireAdmin } from "@/features/auth/lib/auth-server";
+import { requireAdmin } from "@/features/auth/server/lib/auth-server";
 import { invalidateWebCache } from "@/lib/utils/cache-invalidation";
 import type { UpdateTagInput } from "../types";
 import { updateTagRecord } from "../repositories/tag-repository";
